extract traceWorkflowRun into workflow.ts
